# config/media/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rack AI private media API",
    "version": "1.0.0",
    "description": "Bearer principals are server-configured. Browser sessions use HttpOnly cookies and exact Origin checks. Maximum JSON body 16 KiB, 128 active jobs, 2048 retained jobs, 64 HTTP operations and 16 WebSockets. Job admission never requires an interactive session. Retained identities are never silently evicted; full retention returns 503 until deliberate operator archival."
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/api/media/v1/jobs/{job}": {
      "get": {
        "summary": "Read caller-owned job",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Job"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Unknown or another owner",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Identity conflict or terminal conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Body exceeds limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Unsupported or invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Capacity unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "parameters": [
        {
          "name": "job",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ]
    },
...
    "/api/media/v1/sessions/{session}": {
      "get": {
        "summary": "Read caller-owned session",
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Session"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "403": {
            "description": "Not permitted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Unknown or another owner",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Identity conflict or terminal conflict",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "413": {
            "description": "Body exceeds limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Unsupported or invalid request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Capacity unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "parameters": [
        {
          "name": "session",
          "in": "path",
          "required": true,
          "schema": {
            "type": "string",
            "format": "uuid"
          }
        }
      ]
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
...
      "Job": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "schema": {
            "const": "rack-ai/media/v1"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "location": {
            "type": "string"
          },
          "work_id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "submission_id": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "state": {
            "enum": [
              "waiting",
              "starting",
              "dispatching",
              "running",
              "submission_uncertain",
              "completed",
              "failed",
              "cancelled",
              "interrupted"
            ]
          },
          "profile": {
            "type": "string",
            "minLength": 1,
            "maxLength": 128
          },
          "profile_version": {
            "const": 1
          },
          "parameters": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "prompt": {
                "type": "string",
                "minLength": 1,
                "maxLength": 4096
              },
              "negative_prompt": {
                "type": "string",
                "maxLength": 4096
              },
              "seed": {
                "type": "integer",
                "minimum": 0,
                "maximum": 9223372036854775807
              },
              "width": {
                "type": "integer",
                "minimum": 64,
                "maximum": 1024,
                "multipleOf": 64
              },
              "height": {
                "type": "integer",
                "minimum": 64,
                "maximum": 1024,
                "multipleOf": 64
              },
              "steps": {
                "type": "integer",
                "minimum": 1,
                "maximum": 50
              }
            },
            "required": [
              "prompt",
              "negative_prompt",
              "seed",
              "width",
              "height",
              "steps"
            ]
          },
          "prompt_id": {
            "type": "string",
            "format": "uuid"
          },
          "activation": {
            "type": [
              "string",
              "null"
            ]
          },
          "workflow_sha256": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$"
          },
          "model_identity": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$"
          },
          "created_at": {
            "type": "integer",
            "minimum": 0
          },
          "updated_at": {
            "type": "integer",
            "minimum": 0
          },
          "cancel_requested": {
            "type": "boolean"
          },
          "artifacts": {
            "type": "array",
            "maxItems": 1,
            "items": {
              "type": "object",
              "additionalProperties": false,
              "properties": {
                "id": {
                  "type": "string",
                  "format": "uuid"
                },
                "filename": {
                  "const": "image.png"
                },
                "sha256": {
                  "type": "string",
                  "pattern": "^[a-f0-9]{64}$"
                },
                "bytes": {
                  "type": "integer",
                  "minimum": 1,
                  "maximum": 16777216
                },
                "content_type": {
                  "const": "image/png"
                },
                "width": {
                  "type": "integer",
                  "minimum": 64,
                  "maximum": 1024,
                  "multipleOf": 64
                },
                "height": {
                  "type": "integer",
                  "minimum": 64,
                  "maximum": 1024,
                  "multipleOf": 64
                }
              },
              "required": [
                "id",
                "filename",
                "sha256",
                "bytes",
                "content_type",
                "width",
                "height"
              ]
            }
          },
          "error": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "schema",
          "id",
          "location",
          "work_id",
          "submission_id",
          "state",
          "profile",
          "profile_version",
          "parameters",
          "prompt_id",
          "activation",
          "workflow_sha256",
          "model_identity",
          "created_at",
          "updated_at",
          "cancel_requested",
          "artifacts",
          "error"
        ]
      },
...
      "Session": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "schema": {
            "const": "rack-ai/media/v1"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "location": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "release_requested": {
            "type": "boolean"
          },
          "access_url": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "schema",
          "id",
          "location",
          "state",
          "release_requested",
          "access_url"
        ]
      },
      "Error": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "schema": {
            "const": "rack-ai/media/v1"
          },
          "error": {
            "type": "object",
            "additionalProperties": false,
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "schema",
          "error"
        ]
      }
    }
  }
}